move all conversions to type_conversion.rs
